Collapse PipelineBuilder into DrawablePipeline and move pipelinereferences into the drawable module
